+++ test/fixture/simple-private.output.md.json
@@ -1,0 +1,4 @@
+{
+  "type": "root",
+  "children": []
+}
