+++ test/fixture/simple-singlestar.output.md.json
@@ -1,0 +1,4 @@
+{
+  "type": "root",
+  "children": []
+}
