+++ test/fixture/simple-triplestar.output.md.json
@@ -1,0 +1,4 @@
+{
+  "type": "root",
+  "children": []
+}
